+++ src/alsyntaxmodifiers/workspaceCommandSyntaxModifier.ts
@@ -16,7 +16,7 @@ export class WorkspaceCommandSyntaxModifier extends SyntaxModifier {
     }
 
     async runForWorkspaceWithoutUI(workspaceUri: vscode.Uri): Promise<ISyntaxModifierResult | undefined> {
-        let request: ToolsWorkspaceCommandRequest = new ToolsWorkspaceCommandRequest(this._commandName, '', workspaceUri.fsPath, undefined, undefined, this.getParameters(workspaceUri));       
+        let request: ToolsWorkspaceCommandRequest = new ToolsWorkspaceCommandRequest(this._commandName, '', workspaceUri.fsPath, undefined, undefined, this.getParameters(workspaceUri));
         let response = await this.runWorkspaceCommand(request);
 
         if (response) {